refactor: reorganize directory structure
ディレクトリ構成改善のため、root下に`actions/`と`services/`を作り、
関数・モジュールを適切に移動しました。
Fixes #10
