+++ lib/errors.js
@@ -1,4 +1,4 @@
-// lib/shared/errors.js
+// lib/errors.js
 export class HttpError extends Error {
   constructor(status, message = "", opts = {}) {
     super(message || String(status));
