+++ services/channels/ensureHome.js
@@ -1,5 +1,5 @@
-// lib/server/services/channels.js
-import { prisma } from "@/lib/server/prisma";
+// services/channels/ensureHome.js
+import { prisma } from "@/lib/prisma";
 
 const homeSelect = { id: true, slug: true, name: true, isHome: true };
 
@@ -12,13 +12,13 @@ export async function ensureHome({ userId, tx = prisma, homeName }) {
     where: { userId_name: { userId, name: homeName } }, // @@unique([userId, name])
     update: {}, // 既存ならそのまま
     create: { userId, isHome: true, name: homeName }, // なければ作成
-    select: homeSelect
+    select: homeSelect,
   });
 
   // User.homeIdをhome.idに更新（すでに設定済みでも問題ない）
   await tx.user.update({
     where: { id: userId },
-    data: { homeId: home.id }
+    data: { homeId: home.id },
   });
 
   return home;
