+++ services/users/getOrCreateUser.js
@@ -1,7 +1,6 @@
-// lib/server/services/users.js
-import { prisma } from "@/lib/server/prisma";
-import { getUserFromToken } from "@/lib/server/user/getUserFromToken";
-import { HttpError } from "@/lib/shared/errors";
+// services/users/getOrCreateUser.js
+import { prisma } from "@/lib/prisma";
+import { getUserFromToken } from "@/services/users/getUserFromToken";
 
 // userを探して返すが、見つからなければ作成して返す、userの存在を保証する。副作用あり
 // bootstrapUserでのみ呼ばれている（index.jsでのみ呼ばれている）
@@ -16,15 +15,4 @@ export async function getOrCreateUser({ token, tx = prisma }) {
   });
 
   return { user, created: true };
-}
-
-// tokenに対してuserがnullなら外側にエラーを投げて処理中断
-export async function requireUser({ token, tx = prisma }) {
-  const user = await getUserFromToken({ token, tx }); // user || null
-  if (!user)
-    throw new HttpError(401, "Unauthorized", {
-      code: "UNAUTHORIZED",
-      meta: { tokenPresent: Boolean(token) },
-    });
-  return user; // at SERVER_USER_SELECT
 }
